Add test for models of todo and project - install dependencies of uuid and remove cdn link from index.html - Move test files from ./src/scripts to ./test folder
